requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==4.1.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.7.0
pymongo==3.11.4
pytz==2024.2
sqlparse==0.2.4
tzdata==2024.2
